Gate wasmprinter's component support behind a feature
Similar to other recent PRs there's now a `component-model` feature for
the `wasmprinter` crate which is enabled-by-default but can be disabled
if component printing is not desired.
