+++ Cargo.toml
@@ -103,7 +103,7 @@ wasm-mutate = { version = "0.218.0", path = "crates/wasm-mutate" }
 wasm-shrink = { version = "0.218.0", path = "crates/wasm-shrink" }
 wasm-smith = { version = "0.218.0", path = "crates/wasm-smith" }
 wasmparser = { version = "0.218.0", path = "crates/wasmparser", default-features = false, features = ['std'] }
-wasmprinter = { version = "0.218.0", path = "crates/wasmprinter" }
+wasmprinter = { version = "0.218.0", path = "crates/wasmprinter", default-features = false }
 wast = { version = "218.0.0", path = "crates/wast" }
 wat = { version = "1.218.0", path = "crates/wat" }
 wit-component = { version = "0.218.0", path = "crates/wit-component" }
