+++ crates/wasm-encoder/Cargo.toml
@@ -30,7 +30,7 @@ wasmparser = { optional = true, workspace = true }
 anyhow = { workspace = true }
 tempfile = "3.2.0"
 wasmparser = { path = "../wasmparser" }
-wasmprinter = { path = "../wasmprinter" }
+wasmprinter = { workspace = true }
 
 [features]
 default = ['component-model']
