+++ crates/wasm-mutate/Cargo.toml
@@ -21,7 +21,7 @@ egg = "0.6.0"
 
 [dev-dependencies]
 anyhow = { workspace = true }
-wat = { path = "../wat" }
-wasmprinter = { path = "../wasmprinter" }
+wat = { workspace = true }
+wasmprinter = { workspace = true }
 env_logger = { workspace = true }
 wasmparser = { workspace = true, features = ['validate', 'features'] }
