+++ fuzz/Cargo.toml
@@ -17,17 +17,17 @@ env_logger = { workspace = true }
 libfuzzer-sys = { workspace = true }
 log = { workspace = true }
 tempfile = "3.0"
-wasm-mutate = { path = "../crates/wasm-mutate" }
-wasm-smith = { path = "../crates/wasm-smith", features = ['component-model'] }
-wasmparser = { path = "../crates/wasmparser" }
-wasmprinter = { path = "../crates/wasmprinter" }
+wasm-mutate = { workspace = true }
+wasm-smith = { workspace = true, features = ['component-model'] }
+wasmparser = { workspace = true }
+wasmprinter = { workspace = true }
 wasmtime = { workspace = true, optional = true }
-wast = { path = "../crates/wast" }
-wat = { path = "../crates/wat" }
-wit-parser = { path = "../crates/wit-parser" }
-wit-smith = { path = "../crates/wit-smith" }
-wit-component = { path = "../crates/wit-component", features = ['dummy-module'] }
-wasm-encoder = { path = "../crates/wasm-encoder" }
+wast = { workspace = true }
+wat = { workspace = true }
+wit-parser = { workspace = true }
+wit-smith = { workspace = true }
+wit-component = { workspace = true, features = ['dummy-module'] }
+wasm-encoder = { workspace = true }
 
 [dependencies.wit-parser-old]
 package = 'wit-parser'
